extract desktop.erb layout
Other pages have the search bar at the top, but not the jumbotron. Thus,
we need to pull the search bar out of that and give it its own
background color.

Similarly, other pages have the credits footer at the bottom, but not
the stats area. Thus, we need to pull the credits out of that and give
it its own styling.